% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{.addREs}
\alias{.addREs}
\title{Add random effects to the data frame}
\usage{
.addREs(obj, newdata)
}
\arguments{
\item{obj}{object with model fit}

\item{newdata}{an object coercible to \code{data.table}.}
}
\description{
Internal function
}
