% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{fixef.hierCredGLM}
\alias{fixef.hierCredGLM}
\alias{fixef.hierCredTweedie}
\title{Extract fixed-effects estimates}
\usage{
\method{fixef}{hierCredGLM}(object, ...)

\method{fixef}{hierCredTweedie}(object, ...)
}
\arguments{
\item{object}{any fitted model object from which fixed effects estimates can be extracted.}

\item{...}{ignored.}
}
\value{
a named, numeric vector of fixed-effects estimates.
}
\description{
Extract fixed-effects estimates
}
\examples{
\dontrun{
fit = hierCredGLM(Y ~ area + (1 | VehicleType / VehicleBody), dataCar, weights = w, p = 1.75, epsilon = 1e-6)
fixef(fit)
}
}
