% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{hierCredibility-class}
\alias{hierCredibility-class}
\alias{print.hierCredibility}
\alias{summary.hierCredibility}
\alias{fitted.hierCredibility}
\title{Class "hierCredibility" of fitted hierarchical credibility models}
\usage{
\method{print}{hierCredibility}(x, ...)

\method{summary}{hierCredibility}(object, ...)

\method{fitted}{hierCredibility}(object, ...)
}
\arguments{
\item{x}{an object of class \code{\link{hierCredibility}}}

\item{...}{currently ignored.}

\item{object}{an object of class \code{\link{hierCredibility}}}
}
\description{
Class "hierCredibility" of fitted hierarchical credibility models
}
\section{{S3 methods}}{

\describe{
 \item{\code{print}:}{Prints the \code{call}, the estimated variance parameters and the unique number of categories
  of the hierarchical MLF. The \code{...} argument is currently ignored. Returns an invisible copy of the original
  object.}
 \item{\code{summary}:}{In addition to the output of the \code{print.hierCredibility} function, the \code{summary} function
  prints the random effect estimates as well. Returns an invisible copy of the original object.}
  \item{\code{fitted}:}{Returns the fitted values.}
}
}

\seealso{
\code{\link{hierCredibility}}
}
