% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{is.formula}
\alias{is.formula}
\title{Formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{the object.}
}
\value{
logical indicating if the object is a formula or not.
}
\description{
Checks if the object is of the type formula
}
\examples{
f = formula(y ~ x)
is.formula(f)
}
