% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{NrUnique}
\alias{NrUnique}
\title{Number of unique elements in a vector}
\usage{
NrUnique(x, na.rm = TRUE)
}
\arguments{
\item{x}{object of type \code{vector}.}

\item{na.rm}{logical indicating if missing values have to be removed. Default to \code{TRUE}.}
}
\value{
vector with the number of unique elements
}
\description{
Number of unique elements in a vector
}
\examples{
set.seed(1)
x = sample(letters, 50, TRUE)
NrUnique(x)
}
