% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRE.R
\name{plotRE}
\alias{plotRE}
\title{Plotting the random effect estimates}
\usage{
plotRE(obj, levelRE = c("all", "first", "second"), colour = "black", plot = T)
}
\arguments{
\item{obj}{an object of type \code{hierCredibility}, \code{hierCredGLM} or \code{hierCredTweedie}}

\item{levelRE}{indicates which hierarchical level has to be used.}

\item{colour}{colour for \code{geom_point}}

\item{plot}{logical indicating if the \code{ggplot} objects have to be plotted.}
}
\value{
a list with \code{ggplot} objects.
}
\description{
Plotting the random effect estimates
}
\examples{
\dontrun{
fitHGLM <- hierCredGLM(Y ~ area + gender + (1 | VehicleType / VehicleBody), dataCar, weights = w)
plotRE(fitHGLM)
}
}
