% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.hierCredibility}
\alias{predict.hierCredibility}
\title{Model predictions}
\usage{
\method{predict}{hierCredibility}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{...}{ignored.}
}
\value{
If \code{newdata} is omitted the predictions are based on the data used for the fit.
}
\description{
Obtain predictions based on a model fit with \code{hierCredibility}
}
\seealso{
\code{\link{hierCredibility}}
}
