% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.hierCredTweedie}
\alias{predict.hierCredTweedie}
\title{Model predictions}
\usage{
\method{predict}{hierCredTweedie}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{...}{arguments passed to \code{\link[cplm]{cpglm}}}
}
\value{
If \code{newdata} is omitted the predictions are based on the data used for the fit.
}
\description{
Obtain predictions based on the model fit with \code{hierCredTweedie}
}
\details{
The random effects are taken into account by specifying these as an offset in the \code{predict.cpglm} function.
}
\seealso{
\code{\link{hierCredTweedie}}
}
