% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{print.BalanceProperty}
\alias{print.BalanceProperty}
\title{Print method for an object of class \code{BalanceProperty}}
\usage{
\method{print}{BalanceProperty}(x, ...)
}
\arguments{
\item{x}{an object of type \code{BalanceProperty}}

\item{...}{Currently ignored.}
}
\value{
Prints the call and whether the balance property is satisfied or not. Returns an invisible copy
of the original object.
}
\description{
Print method for an object of class \code{BalanceProperty}
}
\seealso{
\code{\link{BalanceProperty}}
}
