% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{ranef-actuaRE}
\alias{ranef-actuaRE}
\alias{ranef.hierCredibility}
\alias{ranef.hierCredGLM}
\alias{ranef.hierCredTweedie}
\title{Extract the random effect estimates from a fitted random effects model}
\usage{
\method{ranef}{hierCredibility}(object, ...)

\method{ranef}{hierCredGLM}(object, ...)

\method{ranef}{hierCredTweedie}(object, ...)
}
\arguments{
\item{object}{an object of type \code{\link{hierCredibility}}, \code{\link{hierCredGLM}} or \code{\link{hierCredTweedie}}}

\item{...}{Currently ignored.}
}
\description{
Extract the random effect estimates from a fitted random effects model
}
