% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{ranef.hierCredGLM}
\alias{ranef.hierCredGLM}
\title{Extract Random Effects}
\usage{
\method{ranef}{hierCredGLM}(object, ...)
}
\arguments{
\item{object}{any fitted model object from which random effects
   estimates can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\value{
will depend on the method function used; see the appropriate documentation.
}
\description{
This function is generic; method functions can be written to handle
  specific classes of objects. Classes which already have methods for
  this function include \code{lmList} and \code{lme}.
}
\examples{
## see the method function documentation
}
\references{
Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models
 in S and S-PLUS", Springer, esp. pp. 100, 461.
}
\seealso{
\code{\link[nlme]{ranef.lmList}},\code{\link[nlme]{ranef.lme}}
}
