% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{BalanceProperty}
\alias{BalanceProperty}
\title{Balance property}
\usage{
BalanceProperty(obj)
}
\arguments{
\item{obj}{an object containing the model fit}
}
\value{
a list with the slots \code{call} (the original call), \code{BalanceProperty} (logical indicating whether the balance
property is satisfied) and \code{Alpha} (Ratio total observed damage to total predicted damage).
}
\description{
Function to assess whether the balance property holds
}
\examples{
\donttest{
fit = hierCredGLM(Y ~ area + (1 | VehicleType / VehicleBody), dataCar, weights = w,
 p = 1.75, epsilon = 1e-6)
BalanceProperty(fit)
}
}
