% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{ranef-actuaRE}
\alias{ranef-actuaRE}
\alias{ranef.hierCredibility}
\alias{ranef.hierCredGLM}
\alias{ranef.hierCredTweedie}
\title{Extract the random effect estimates from a fitted random effects model}
\usage{
\method{ranef}{hierCredibility}(object, ...)

\method{ranef}{hierCredGLM}(object, ...)

\method{ranef}{hierCredTweedie}(object, ...)
}
\arguments{
\item{object}{an object of type \code{\link{hierCredibility}}, \code{\link{hierCredGLM}} or \code{\link{hierCredTweedie}}}

\item{...}{Currently ignored.}
}
\value{
A list of data frames, one for each grouping factor for the random effects. The number of rows in the data frame is the number of levels of the grouping factor.
The first (two) columns correspond(s) to the grouping factor. The last column corresponds to the estimated random effect.
}
\description{
A generic function to extract the estimates/predictions of the random effects from a fitted random effects model.
}
